+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.2
+
+* Add Swift Package Manager support for iOS.
+* Update Flutter and Dart SDK constraints for current Flutter tooling.
+
 ## 1.1.1
 
 * Update code